+++ Editor/CSV/CsvDiff.cs
@@ -1,3 +1,4 @@
+#if UNITY_EDITOR
 using System;
 using System.Collections.Generic;
 
@@ -221,3 +222,4 @@ namespace TRnK.Localization
         }
     }
 }
+#endif