+++ Editor/CSV/CsvParser.cs
@@ -1,3 +1,4 @@
+#if UNITY_EDITOR
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -145,3 +146,4 @@ namespace TRnK.Localization
         }
     }
 }
+#endif