+++ Editor/CSV/CsvWriter.cs
@@ -1,3 +1,4 @@
+#if UNITY_EDITOR
 using System.Collections.Generic;
 using System.Text;
 
@@ -70,3 +71,4 @@ namespace TRnK.Localization
         }
     }
 }
+#endif